# docs/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cloacina-server API",
    "description": "REST API for the Cloacina workflow-orchestration server. WebSocket protocols (execution events, accumulator ingest, reactor commands, substrate delivery) are specified in the WebSocket Protocol reference of the documentation site.",
    "contact": {
      "name": "Dylan Storey",
      "email": "dylan.storey@gmail.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.7.0"
  },
  "paths": {
...
    "/v1/tenants/{tenant_id}/triggers": {
      "get": {
        "tags": [
          "triggers"
        ],
        "summary": "GET /tenants/:tenant_id/triggers — list all schedules (cron + trigger).",
        "description": "CLOACI-T-0579: routed through the tenant-scoped `Database` from\n`TenantDatabaseCache` so the underlying `SELECT FROM schedules`\nhits the tenant's schema, not the admin schema. Closes SEC-02.",
        "operationId": "list_triggers",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedules page (cron + trigger)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantListResponse_TriggerScheduleSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/v1/tenants/{tenant_id}/triggers/{name}": {
      "get": {
        "tags": [
          "triggers"
        ],
        "summary": "GET /tenants/:tenant_id/triggers/:name — trigger details + recent executions.",
        "description": "CLOACI-T-0579: routed through the tenant-scoped `Database`. A request\nfor tenant B with a trigger id that belongs to tenant A naturally\n404s — the row simply doesn't exist in tenant B's schema. No\ninfo-disclosure via \"not in your tenant\" error code. Closes SEC-02.",
        "operationId": "get_trigger",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Trigger or workflow name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trigger detail + recent executions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Trigger not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Standardized error response body. Every non-2xx response from\n`cloacina-server` carries this shape; the request correlation ID is in\nthe `x-request-id` response header, not the body.",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code (e.g. `invalid_pagination`,\n`workflow_not_found`)."
          },
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "TenantListResponse_TriggerScheduleSummary": {
        "type": "object",
        "description": "List envelope variant that retains a top-level `tenant_id`, used by\ntenant-scoped list endpoints for backward compatibility with operator\ndashboards that key off it.",
        "required": [
          "tenant_id",
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "One row in the trigger list.",
              "required": [
                "id",
                "schedule_type",
                "workflow_name",
                "enabled",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "description": "RFC 3339 timestamp."
                },
                "cron_expression": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "enabled": {
                  "type": "boolean"
                },
                "id": {
                  "type": "string",
                  "description": "Schedule UUID."
                },
                "last_run_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "RFC 3339 timestamp."
                },
                "next_run_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "RFC 3339 timestamp."
                },
                "poll_interval_ms": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "schedule_type": {
                  "type": "string",
                  "description": "`cron` or `trigger`."
                },
                "trigger_name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "workflow_name": {
                  "type": "string"
                }
              }
            }
          },
          "tenant_id": {
            "type": "string"
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "TriggerDetailResponse": {
        "type": "object",
        "description": "`GET /tenants/{tenant_id}/triggers/{name}` response.",
        "required": [
          "tenant_id",
          "schedule",
          "recent_executions"
        ],
        "properties": {
          "recent_executions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TriggerExecution"
            }
          },
          "schedule": {
            "$ref": "#/components/schemas/TriggerScheduleInfo"
          },
          "tenant_id": {
            "type": "string"
          }
        }
      },
      "TriggerExecution": {
        "type": "object",
        "description": "One row in `recent_executions` of the trigger detail response.",
        "required": [
          "id",
          "started_at"
        ],
        "properties": {
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339 timestamp."
          },
          "id": {
            "type": "string",
            "description": "Schedule-execution UUID."
          },
          "scheduled_time": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339 timestamp."
          },
          "started_at": {
            "type": "string",
            "description": "RFC 3339 timestamp."
          }
        }
      },
      "TriggerScheduleInfo": {
        "type": "object",
        "description": "Schedule fields in the trigger detail response.",
        "required": [
          "id",
          "schedule_type",
          "workflow_name",
          "enabled"
        ],
        "properties": {
          "cron_expression": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "description": "Schedule UUID."
          },
          "schedule_type": {
            "type": "string",
            "description": "`cron` or `trigger`."
          },
          "trigger_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cloacina API key, sent as `Authorization: Bearer <key>`. Tenant scope is carried by the key itself; tenant-scoped routes additionally name the tenant in the path."
      }
    }
  },
...
}